third_party/WebKit/LayoutTests/inspector/console/console-format-es6-symbols-error.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/console-test.js"></script>
<script>

function logSymbolToConsoleWithError()
{
    Symbol.prototype.toString = function (arg) { throw new Error(); };
    var smb = Symbol();
    console.log(smb);
}

function test()
{
    InspectorTest.evaluateInPage("logSymbolToConsoleWithError()", complete);

    function complete()
    {
        InspectorTest.dumpConsoleMessages();
        InspectorTest.completeTest();
    }
}
</script>
</head>

<body onload="runTest()">
<p>
Tests that console properly displays information about ES6 Symbols.
</p>
</body>
</html>
